better for epoc channel data now
changed the channel processing so that each could have a specific max
value based on where they sit on their average value. but that is
somthing for later if there is time.

also made a backup of the file the way it was. even though that is what
revision control is for.